adote-pet-back/requirements.txt
Django>=4.0
djangorestframework>=3.13
django-cors-headers>=3.13
mysqlclient>=2.0
